+++ sample/src/androidMain/kotlin/com/github/panpf/sketch/sample/ui/gallery/PhotoGridItemFactory.kt
@@ -110,6 +110,7 @@ class PhotoGridItemFactory constructor(val animatedPlaceholder: Boolean = false)
                 }
                 crossfade()
                 resizeOnDraw()
+                sizeMultiplier(2f)  // To get a clearer thumbnail
             }
         }
     }