+++ sample/src/commonMain/kotlin/com/github/panpf/sketch/sample/ui/gallery/PhotoGridItem.kt
@@ -116,6 +116,7 @@ fun PhotoGridItem(
             }
             crossfade()
             resizeOnDraw()
+            sizeMultiplier(2f)  // To get a clearer thumbnail
             merge(appSettingsService.buildListImageOptions())
         }
     }