+++ sketch-test-core/src/main/kotlin/com/github/panpf/sketch/test/utils/ImageRequestUtils.kt
@@ -10,13 +10,14 @@ import com.github.panpf.sketch.request.internal.RequestDelegate
 import com.github.panpf.sketch.request.internal.RequestManager
 import com.github.panpf.sketch.target.Target
 import com.github.panpf.sketch.util.Size
+import com.github.panpf.sketch.util.times
 import kotlinx.coroutines.Job
 import kotlinx.coroutines.runBlocking
 
 
 fun ImageRequest.toRequestContext(sketch: Sketch, size: Size? = null): RequestContext {
     return RequestContext(sketch, this).apply {
-        this@apply.size = size ?: runBlocking { sizeResolver.size() }
+        this@apply.size = size ?: runBlocking { sizeResolver.size() * (sizeMultiplier ?: 1f) }
     }
 }